+++ common/type_system/TypeSystem.h
@@ -104,6 +104,8 @@ struct FieldReverseLookupOutput {
   double total_score = 0.;
   TypeSpec result_type;
   std::vector<Token> tokens;
+
+  bool has_variable_token() const;
 };
 
 struct FieldReverseMultiLookupOutput {
@@ -159,12 +161,13 @@ class TypeSystem {
   MethodInfo declare_method(const std::string& type_name,
                             const std::string& method_name,
                             bool no_virtual,
-                            const TypeSpec& ts);
+                            const TypeSpec& ts,
+                            bool override_type);
   MethodInfo declare_method(Type* type,
                             const std::string& method_name,
                             bool no_virtual,
-                            const TypeSpec& ts);
-
+                            const TypeSpec& ts,
+                            bool override_type);
   MethodInfo define_method(const std::string& type_name,
                            const std::string& method_name,
                            const TypeSpec& ts);