+++ decompiler/ObjectFile/LinkedObjectFile.h
@@ -40,6 +40,7 @@ class LinkedObjectFile {
   void symbol_link_offset(int source_segment, int source_offset, const char* name);
   Function& get_function_at_label(int label_id);
   const Function* try_get_function_at_label(int label_id) const;
+  const Function* try_get_function_at_label(const DecompilerLabel& label) const;
   std::string get_label_name(int label_id) const;
   uint32_t set_ordered_label_names();
   void find_code();